refactored code to have API call in <App>. Cannot figure out how to get the :id of the url that <Route> is set up for.  not understanding how to user withRouter() in this context.  Onlly able to get the props.match.params of the routed component
