+++ client/src/App.js
@@ -1,17 +1,38 @@
-import React, { Component } from 'react';
-import { Route, Link } from 'react-router-dom';
+import React from 'react';
+import { Route } from 'react-router-dom';
+import axios from 'axios';
 import SavedList from './Movies/SavedList';
-// import MovieList from './Movies/MovieList';
-// import Movie from './Movies/Movie';
+import MovieList from './Movies/MovieList';
 import MovieCard from './Movies/MovieCard';
+import {withRouter} from 'react-router-dom';
 
-export default class App extends Component {
+
+
+class App extends React.Component {
   constructor() {
     super();
     this.state = {
-      savedList: []
+      savedList: [],
+      movies: []
     };
   }
+
+  componentDidMount() { 
+    // const id = this.props.match.params.id;
+    const id = this.state.match;
+    this.fetchMovie(id);
+  }
+
+  fetchMovie = () => {
+    axios 
+      .get(`http://localhost:5000/api/movies/`)
+      .then(response => {
+        this.setState(() => ({ movies: response.data }));
+      })
+      .catch(error => {
+        console.error('Server Error', error);
+      });
+  };
 
   addToSavedList = movie => {
     const savedList = this.state.savedList;
@@ -23,10 +44,11 @@ export default class App extends Component {
     return (
         <div>
           <SavedList list={this.state.savedList} />
-          {/* <Route path='/' exact component={MovieList} /> */}
-          <Route path='/movies/:id' component={MovieCard} />
+          <Route path='/' exact render={ (props) => <MovieList {...props} movies={this.state.movies} /> } />
+          <Route path='/movies/:id' render={ (props) => <MovieCard {...props} movies={this.state.movies} /> } />
         </div>
     );
   }
 }
 
+export default App;
