+++ client/src/Movies/MovieList.js
@@ -1,57 +1,38 @@
 import React, { Component } from 'react';
-import axios from 'axios';
-import { Link } from 'react-router-dom';
 import './movies.css';
+import MovieDetails from './MovieDetails';
 
-export default class MovieList extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      movies: []
-    };
-    
-  }
-
-  componentDidMount() {
-    axios
-      .get('http://localhost:5000/api/movies')
-      .then(response => {
-        this.setState(() => ({ movies: response.data }));
-      })
-      .catch(error => {
-        console.error('Server Error', error);
-      });
-  }
-
-  render() {
+const MovieList = props => {
+    console.log(props);
     return (
       <div className="movie-list">
-        {this.state.movies.map( movie => (
+        {props.movies.map( movie => (
           <MovieDetails key={movie.id} movie={movie} />
         ))}
       </div>
     );
   }
-}
 
-function MovieDetails({ movie }) {
-  const { title, director, metascore, stars } = movie;
-  return (
-    <div className="movie-card">
-      <Link to={`/movies/${movie.id}`}><button><h2>{title}</h2></button></Link>
-      <div className="movie-director">
-        Director: <em>{director}</em>
-      </div>
-      <div className="movie-metascore">
-        Metascore: <strong>{metascore}</strong>
-      </div>
-      <h3>Actors</h3>
+  export default MovieList;
 
-      {stars.map(star => (
-        <div key={star} className="movie-star">
-          {star}
-        </div>
-      ))}
-    </div>
-  );
-}
+// function MovieDetails({ movie }) {
+//   const { title, director, metascore, stars } = movie;
+//   return (
+//     <div className="movie-card">
+//       <Link to={`/movies/${movie.id}`}><button><h2>{title}</h2></button></Link>
+//       <div className="movie-director">
+//         Director: <em>{director}</em>
+//       </div>
+//       <div className="movie-metascore">
+//         Metascore: <strong>{metascore}</strong>
+//       </div>
+//       <h3>Actors</h3>
+
+//       {stars.map(star => (
+//         <div key={star} className="movie-star">
+//           {star}
+//         </div>
+//       ))}
+//     </div>
+//   );
+// }
